Setup the basics of the localtime gem.
Currently this still requires you to put “datejs-rails” in your project and add the following require lines to your application.js:

//= require date
//= require date/extras
//= require localtime

I don’t think we can solve the problem of having to add requires to the application.js file, but I would like to figure out how daters-rails can be required as a dependency of this gem, rather than having to have it in the actual project’s Gemfile. I’m pretty sure it’s possible, but this code is the bare minimum that could work so I’m committing it as such.